backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv>=0.19.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
python-dateutil==2.8.2
transformers==4.36.2
torch==2.2.0
meteostat==1.6.5 